fix(db): align types and legacy db helpers for tests
- Use BIGINT for mmsi/time in Postgres/Timescale tables
- Make aggregate_static_msgs signature backward compatible
- Allow month-aware drop/rebuild/dedup helpers
- Update TimescaleDB test expectations for global hypertables
- Avoid passing months to aggregate_static_msgs in test helper
- Skip DB setup/tests on Windows runners

+++ aisdb/aisdb_sql/psql_createtable_dynamic_noindex.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS ais_{0}_dynamic (
-    mmsi INTEGER NOT NULL,
-    time INTEGER NOT NULL,
+    mmsi BIGINT NOT NULL,
+    time BIGINT NOT NULL,
     longitude REAL NOT NULL,
     latitude REAL NOT NULL,
     rot REAL,
